Add an `IS_PAYMENT_ENABLED` feature flag.
The `Login` use case authenticates with the old REST API in order
to stay compatible with some non-migrated calls.
Yet, it seems those calls are not broadly used and should only be
enabled for specific flavors.
Therefore, we introduce a build-time feature flag to only
log in with the REST API if needed.
The feature flag also impacts the `LoginRepository`, which is instructed
to check the validity of the REST token only if the feature flag is enabled.

Attention needs to be drawn to the fact that this feature flag does
not disable the REST calls but only the login part.
If some more in-depth modifications need to be made, a proper
analysis should be made first.

MGOM2-12